Adding threadpool

+++ Lib.TradingSystemClassLibrary/Interfaces/ITradeProcessor.cs
@@ -4,9 +4,9 @@ namespace TradingSystem.Interfaces
 {
     public interface ITradeProcessor
     {
-        public Task<int> PlaceOrder(int userId, OrderType orderType, string stockSymbol, int quantity, decimal price);
+        public Task<int> PlaceOrder(int userId, OrderType orderType, string stockSymbol, int quantity, int price);
 
-        public Task<bool> ModifyOrder(int orderId, int quantity, decimal price);
+        public Task<bool> ModifyOrder(int orderId, int quantity, int price);
 
         public Task<bool> CancelOrder(int orderId);
